수정사항 5(02/13 1445i) - 상세 페이지 전기 추가 + 패키지 이름 변경2(#12)
