+++ src/view/AddStationDialog.java
@@ -1,29 +1,33 @@
-package View;
+package view;
 
 import javax.swing.*;
 import javax.swing.border.*;
 import java.awt.*;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
+import java.time.LocalDate;
 
-public class EditProfileDialog extends JDialog {
-    private JTextField nameF, emailF, carF, mileageF;
-    private JRadioButton[] fuelRadios;
-    private ButtonGroup fuelGroup;
+/**
+ * [Add Record Dialog]
+ * 1. 기능 설명: 사용자가 주유 날짜, 장소, 가격, 주유량, 주행거리를 입력하여 개인 차량 관리 데이터를 생성합니다.
+ * 2. API 포인트: [없음] (현재 수동 입력 방식)
+ * 3. DB 포인트: 입력된 정보를 기반으로 fuel_records 테이블에 새로운 행을 INSERT 합니다.
+ * 4. 추가해야 되는 기능: 주유소 명칭 입력 시 인근 주유소 리스트를 추천해주는 API 연동 기능.
+ */
+public class AddStationDialog extends JDialog {
+    private JTextField dateF, stationF, priceF, litersF, mileageF; // priceF 추가
     private JButton saveBtn, cancelBtn;
 
-    public EditProfileDialog(Frame parent) {
-        super(parent, "정보 수정", true);
+    public AddStationDialog(Frame parent) {
+        super(parent, "주유 기록 추가", true);
         setLayout(new BorderLayout());
         setResizable(false);
-        setSize(420, 680);
+        setSize(420, 620);
 
-        /* ===== 전체 배경 ===== */
         JPanel background = new JPanel(new BorderLayout());
         background.setBackground(new Color(243, 244, 246));
         background.setBorder(new EmptyBorder(20, 20, 20, 20));
 
-        /* ===== 카드 패널 ===== */
         JPanel card = new JPanel();
         card.setLayout(new BoxLayout(card, BoxLayout.Y_AXIS));
         card.setBackground(Color.WHITE);
@@ -33,7 +37,7 @@ public class EditProfileDialog extends JDialog {
         ));
         card.setAlignmentX(Component.CENTER_ALIGNMENT);
 
-        /* ===== 상단 헤더 (우측 종료 버튼) ===== */
+        /* ===== 헤더 영역 ===== */
         JPanel header = new JPanel(new BorderLayout());
         header.setBackground(Color.WHITE);
         header.setMaximumSize(new Dimension(Integer.MAX_VALUE, 30));
@@ -53,16 +57,7 @@ public class EditProfileDialog extends JDialog {
         });
         header.add(closeLabel, BorderLayout.EAST);
 
-        /* ===== [DB 포인트 1] 초기 데이터 불러오기 ===== */
-        // 예시 데이터: 실제로는 DB에서 조회한 값을 변수에 할당하세요.
-        String currentName = "홍길동";
-        String currentEmail = "hong@example.com";
-        String currentCar = "12가 3456";
-        String currentFuel = "휘발유";
-        String currentMileage = "50000";
-
-        /* ===== 제목 섹션 ===== */
-        JLabel titleLabel = new JLabel("정보 수정");
+        JLabel titleLabel = new JLabel("주유 기록 추가");
         titleLabel.setFont(new Font(Font.SANS_SERIF, Font.BOLD, 20));
         titleLabel.setAlignmentX(Component.CENTER_ALIGNMENT);
 
@@ -76,27 +71,18 @@ public class EditProfileDialog extends JDialog {
         Font labelFont = new Font(Font.SANS_SERIF, Font.PLAIN, 12);
         Dimension fieldSize = new Dimension(Integer.MAX_VALUE, 35);
 
-        addInput(formWrapper, "이름", nameF = new JTextField(currentName), labelFont, fieldSize);
-        addInput(formWrapper, "이메일", emailF = new JTextField(currentEmail), labelFont, fieldSize);
-        addInput(formWrapper, "차량번호", carF = new JTextField(currentCar), labelFont, fieldSize);
-
-        // 연료 타입 섹션
-        JLabel fuelLabel = new JLabel("연료 타입");
-        fuelLabel.setFont(labelFont);
-        fuelLabel.setAlignmentX(Component.LEFT_ALIGNMENT);
-        formWrapper.add(fuelLabel);
-        formWrapper.add(Box.createVerticalStrut(6));
+        // [기능 설명] 오늘 날짜 자동 세팅
+        addInput(formWrapper, "날짜", dateF = new JTextField(LocalDate.now().toString()), labelFont, fieldSize);
+        addInput(formWrapper, "주유소", stationF = new JTextField(), labelFont, fieldSize);
         
-        JPanel fuelPanel = createFuelPanel(currentFuel);
-        fuelPanel.setAlignmentX(Component.LEFT_ALIGNMENT);
-        formWrapper.add(fuelPanel);
-        formWrapper.add(Box.createVerticalStrut(14));
+        // [추가해야 되는 기능] 현재 위치 기준 해당 주유소의 실시간 가격 API를 호출하여 자동으로 채워주는 기능
+        addInput(formWrapper, "가격 (원)", priceF = new JTextField(), labelFont, fieldSize);
+        
+        addInput(formWrapper, "주유량 (L)", litersF = new JTextField(), labelFont, fieldSize);
+        addInput(formWrapper, "누적 주행 거리 (km)", mileageF = new JTextField(), labelFont, fieldSize);
 
-        // 주행 거리 필드
-        addInput(formWrapper, "주행 거리 (km)", mileageF = new JTextField(currentMileage), labelFont, fieldSize);
-
-        /* ===== 하단 버튼  ===== */
-        saveBtn = new JButton("저장");
+        /* ===== 저장 버튼 및 로직 ===== */
+        saveBtn = new JButton("추가");
         saveBtn.setBackground(new Color(37, 99, 235));
         saveBtn.setForeground(Color.WHITE);
         saveBtn.setFont(new Font(Font.SANS_SERIF, Font.BOLD, 15));
@@ -106,9 +92,24 @@ public class EditProfileDialog extends JDialog {
         saveBtn.setMaximumSize(new Dimension(320, 45));
 
         saveBtn.addActionListener(e -> {
-            /* ===== [DB 포인트 2] 데이터 업데이트 ===== */
-            JOptionPane.showMessageDialog(this, "성공적으로 수정되었습니다.");
-            dispose();
+            try {
+                // [기능 설명] 필드 값 추출 및 숫자 데이터 파싱 검증
+                String date = dateF.getText();
+                String station = stationF.getText();
+                int price = Integer.parseInt(priceF.getText()); // 가격 추출
+                double liters = Double.parseDouble(litersF.getText());
+                int mileage = Integer.parseInt(mileageF.getText());
+
+                // [DB 포인트] 사용자가 입력한 필드 데이터를 쿼리 파라미터로 매핑하여 DB에 저장
+                // SQL: INSERT INTO fuel_records (user_id, log_date, station_name, fuel_price, liters, mileage) 
+                //      VALUES (?, ?, ?, ?, ?, ?);
+
+                JOptionPane.showMessageDialog(this, "주유 기록이 성공적으로 추가되었습니다.");
+                dispose();
+            } catch (NumberFormatException ex) {
+                // [추가해야 되는 기능] 어떤 필드가 잘못되었는지 사용자에게 구체적으로 알려주는 포커스 기능
+                JOptionPane.showMessageDialog(this, "가격, 리터, 주행거리는 숫자만 입력 가능합니다.", "입력 오류", JOptionPane.ERROR_MESSAGE);
+            }
         });
 
         cancelBtn = new JButton("취소");
@@ -120,13 +121,13 @@ public class EditProfileDialog extends JDialog {
         cancelBtn.setAlignmentX(Component.CENTER_ALIGNMENT);
         cancelBtn.addActionListener(e -> dispose());
 
-        /* ===== 카드 조립 ===== */
+        /* ===== 컴포넌트 조립 ===== */
         card.add(header);
         card.add(Box.createVerticalStrut(10));
         card.add(titleLabel);
         card.add(Box.createVerticalStrut(25));
         card.add(formWrapper);
-        card.add(Box.createVerticalStrut(10));
+        card.add(Box.createVerticalStrut(15));
         card.add(saveBtn);
         card.add(Box.createVerticalStrut(12));
         card.add(cancelBtn);
@@ -139,32 +140,18 @@ public class EditProfileDialog extends JDialog {
     private void addInput(JPanel p, String title, JTextField tf, Font font, Dimension size) {
         JLabel lbl = new JLabel(title);
         lbl.setFont(font);
+        lbl.setForeground(new Color(55, 65, 81));
         lbl.setAlignmentX(Component.LEFT_ALIGNMENT);
         p.add(lbl);
-        p.add(Box.createVerticalStrut(6));
+        p.add(Box.createVerticalStrut(5));
 
         tf.setMaximumSize(size);
         tf.setAlignmentX(Component.LEFT_ALIGNMENT);
+        tf.setBorder(new CompoundBorder(
+            new LineBorder(new Color(209, 213, 219), 1),
+            new EmptyBorder(0, 10, 0, 10)
+        ));
         p.add(tf);
-        p.add(Box.createVerticalStrut(14));
-    }
-
-    private JPanel createFuelPanel(String selected) {
-        JPanel p = new JPanel(new GridLayout(2, 3, 5, 5));
-        p.setBackground(Color.WHITE);
-        p.setMaximumSize(new Dimension(320, 55));
-        String[] fuels = {"휘발유", "경유", "LPG", "전기", "하이브리드", "기타"};
-        fuelRadios = new JRadioButton[fuels.length];
-        fuelGroup = new ButtonGroup();
-
-        for (int i = 0; i < fuels.length; i++) {
-            fuelRadios[i] = new JRadioButton(fuels[i]);
-            fuelRadios[i].setBackground(Color.WHITE);
-            fuelRadios[i].setFont(new Font(Font.SANS_SERIF, Font.PLAIN, 12));
-            if (fuels[i].equals(selected)) fuelRadios[i].setSelected(true);
-            fuelGroup.add(fuelRadios[i]);
-            p.add(fuelRadios[i]);
-        }
-        return p;
+        p.add(Box.createVerticalStrut(12));
     }
 }
