+++ src/view/EditProfileDialog.java
@@ -1,4 +1,4 @@
-package View;
+package view;
 
 import javax.swing.*;
 import javax.swing.border.*;
@@ -6,23 +6,19 @@ import java.awt.*;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
 
-/**
- * [Password Change Dialog]
- * 핵심 역할:
- * 1. 입력 검증: 현재 PW 확인, 새 PW와 확인 PW의 일치 여부 체크
- * 2. DB 업데이트: 비밀번호 컬럼 UPDATE (암호화 권장)
- */
-public class PasswordChangeDialog extends JDialog {
-    private JPasswordField currentPwF, newPwF, confirmPwF;
+public class EditProfileDialog extends JDialog {
+    private JTextField nameF, emailF, carF, mileageF;
+    private JRadioButton[] fuelRadios;
+    private ButtonGroup fuelGroup;
     private JButton saveBtn, cancelBtn;
 
-    public PasswordChangeDialog(Frame parent) {
-        super(parent, "비밀번호 변경", true);
+    public EditProfileDialog(Frame parent) {
+        super(parent, "정보 수정", true);
         setLayout(new BorderLayout());
         setResizable(false);
-        setSize(420, 520);
+        setSize(420, 680);
 
-        /* ===== 전체 배경 (메인 톤 유지) ===== */
+        /* ===== 전체 배경 ===== */
         JPanel background = new JPanel(new BorderLayout());
         background.setBackground(new Color(243, 244, 246));
         background.setBorder(new EmptyBorder(20, 20, 20, 20));
@@ -57,26 +53,49 @@ public class PasswordChangeDialog extends JDialog {
         });
         header.add(closeLabel, BorderLayout.EAST);
 
+        /* ===== [DB 포인트 1] 초기 데이터 불러오기 ===== */
+        // 예시 데이터: 실제로는 DB에서 조회한 값을 변수에 할당하세요.
+        String currentName = "홍길동";
+        String currentEmail = "hong@example.com";
+        String currentCar = "12가 3456";
+        String currentFuel = "휘발유";
+        String currentMileage = "50000";
+
         /* ===== 제목 섹션 ===== */
-        JLabel titleLabel = new JLabel("비밀번호 변경");
+        JLabel titleLabel = new JLabel("정보 수정");
         titleLabel.setFont(new Font(Font.SANS_SERIF, Font.BOLD, 20));
         titleLabel.setAlignmentX(Component.CENTER_ALIGNMENT);
 
-        /* ===== 입력 폼 영역 (JPasswordField) ===== */
+        /* ===== 입력 폼 영역 ===== */
         JPanel formWrapper = new JPanel();
         formWrapper.setLayout(new BoxLayout(formWrapper, BoxLayout.Y_AXIS));
         formWrapper.setBackground(Color.WHITE);
         formWrapper.setAlignmentX(Component.CENTER_ALIGNMENT);
-        formWrapper.setMaximumSize(new Dimension(320, 260));
+        formWrapper.setMaximumSize(new Dimension(320, 400));
 
         Font labelFont = new Font(Font.SANS_SERIF, Font.PLAIN, 12);
         Dimension fieldSize = new Dimension(Integer.MAX_VALUE, 35);
 
-        addPasswordInput(formWrapper, "현재 비밀번호", currentPwF = new JPasswordField(), labelFont, fieldSize);
-        addPasswordInput(formWrapper, "새 비밀번호", newPwF = new JPasswordField(), labelFont, fieldSize);
-        addPasswordInput(formWrapper, "새 비밀번호 확인", confirmPwF = new JPasswordField(), labelFont, fieldSize);
+        addInput(formWrapper, "이름", nameF = new JTextField(currentName), labelFont, fieldSize);
+        addInput(formWrapper, "이메일", emailF = new JTextField(currentEmail), labelFont, fieldSize);
+        addInput(formWrapper, "차량번호", carF = new JTextField(currentCar), labelFont, fieldSize);
 
-        /* ===== 하단 버튼 및 로직 ===== */
+        // 연료 타입 섹션
+        JLabel fuelLabel = new JLabel("연료 타입");
+        fuelLabel.setFont(labelFont);
+        fuelLabel.setAlignmentX(Component.LEFT_ALIGNMENT);
+        formWrapper.add(fuelLabel);
+        formWrapper.add(Box.createVerticalStrut(6));
+        
+        JPanel fuelPanel = createFuelPanel(currentFuel);
+        fuelPanel.setAlignmentX(Component.LEFT_ALIGNMENT);
+        formWrapper.add(fuelPanel);
+        formWrapper.add(Box.createVerticalStrut(14));
+
+        // 주행 거리 필드
+        addInput(formWrapper, "주행 거리 (km)", mileageF = new JTextField(currentMileage), labelFont, fieldSize);
+
+        /* ===== 하단 버튼  ===== */
         saveBtn = new JButton("저장");
         saveBtn.setBackground(new Color(37, 99, 235));
         saveBtn.setForeground(Color.WHITE);
@@ -87,29 +106,8 @@ public class PasswordChangeDialog extends JDialog {
         saveBtn.setMaximumSize(new Dimension(320, 45));
 
         saveBtn.addActionListener(e -> {
-            String currentPw = new String(currentPwF.getPassword());
-            String newPw = new String(newPwF.getPassword());
-            String confirmPw = new String(confirmPwF.getPassword());
-
-            // 1. [검증] 빈 칸 체크
-            if (currentPw.isEmpty() || newPw.isEmpty()) {
-                JOptionPane.showMessageDialog(this, "비밀번호를 입력해주세요.", "입력 오류", JOptionPane.WARNING_MESSAGE);
-                return;
-            }
-
-            // 2. [DB Point 1] 현재 비밀번호 대조
-            // SELECT user_pw FROM members WHERE user_id = ?
-            // DB에서 가져온 비밀번호와 currentPw가 일치하는지 먼저 확인해야 합니다.
-
-            // 3. [검증] 새 비밀번호 일치 여부
-            if (!newPw.equals(confirmPw)) {
-                JOptionPane.showMessageDialog(this, "새 비밀번호가 일치하지 않습니다.", "오류", JOptionPane.ERROR_MESSAGE);
-                return;
-            }
-
-            // 4. [DB Point 2] 최종 업데이트
-            // UPDATE members SET user_pw = ? WHERE user_id = ?
-            JOptionPane.showMessageDialog(this, "비밀번호가 성공적으로 변경되었습니다.");
+            /* ===== [DB 포인트 2] 데이터 업데이트 ===== */
+            JOptionPane.showMessageDialog(this, "성공적으로 수정되었습니다.");
             dispose();
         });
 
@@ -138,17 +136,35 @@ public class PasswordChangeDialog extends JDialog {
         setLocationRelativeTo(parent);
     }
 
-    private void addPasswordInput(JPanel p, String title, JPasswordField pf, Font font, Dimension size) {
+    private void addInput(JPanel p, String title, JTextField tf, Font font, Dimension size) {
         JLabel lbl = new JLabel(title);
         lbl.setFont(font);
         lbl.setAlignmentX(Component.LEFT_ALIGNMENT);
         p.add(lbl);
         p.add(Box.createVerticalStrut(6));
 
-        pf.setMaximumSize(size);
-        pf.setAlignmentX(Component.LEFT_ALIGNMENT);
-        pf.setBorder(new CompoundBorder(new LineBorder(new Color(209, 213, 219)), new EmptyBorder(0, 10, 0, 10)));
-        p.add(pf);
+        tf.setMaximumSize(size);
+        tf.setAlignmentX(Component.LEFT_ALIGNMENT);
+        p.add(tf);
         p.add(Box.createVerticalStrut(14));
+    }
+
+    private JPanel createFuelPanel(String selected) {
+        JPanel p = new JPanel(new GridLayout(2, 3, 5, 5));
+        p.setBackground(Color.WHITE);
+        p.setMaximumSize(new Dimension(320, 55));
+        String[] fuels = {"휘발유", "경유", "LPG", "전기", "하이브리드", "기타"};
+        fuelRadios = new JRadioButton[fuels.length];
+        fuelGroup = new ButtonGroup();
+
+        for (int i = 0; i < fuels.length; i++) {
+            fuelRadios[i] = new JRadioButton(fuels[i]);
+            fuelRadios[i].setBackground(Color.WHITE);
+            fuelRadios[i].setFont(new Font(Font.SANS_SERIF, Font.PLAIN, 12));
+            if (fuels[i].equals(selected)) fuelRadios[i].setSelected(true);
+            fuelGroup.add(fuelRadios[i]);
+            p.add(fuelRadios[i]);
+        }
+        return p;
     }
 }
