+++ src/view/RepairPage.java
@@ -1,4 +1,4 @@
-package View;
+package view;
 
 import javax.swing.*;
 import javax.swing.border.*;
